Rename ImageHandle to image overlay
