+++ ts/editor/image-overlay/ImageOverlay.svelte
