+++ ts/editor/image-overlay/index.ts
@@ -1,6 +1,6 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
-import ImageHandle from "./ImageHandle.svelte";
+import ImageOverlay from "./ImageOverlay.svelte";
 
-export default ImageHandle;
+export default ImageOverlay;
